Restructured all the controllers. JS fixes
(draggable building size fix)
