+++ server/conf/routes
@@ -14,7 +14,7 @@ POST        /api/user/auth/admin/migration/run                                  
 
 ### POSITION
 # TODO:NN estimate_position (implement this as well..)
-POST        /anyplace/position/estimate_position                                    controllers.MappingController.findPosition()
+POST        /anyplace/position/estimate_position                                    controllers.PositioningController.estimatePosition()
 
 ###################################
 
@@ -38,14 +38,16 @@ GET /viewer/*file                                                               
 GET /viewer/                                                                        controllers.WebAppController.serveViewer(file="index.html")
 GET /viewer                                                                         controllers.WebAppController.AddTrailingSlash()
 
+
+# TODO:NN put in api.routes as well!!
 ## BLUEPRINTS
-POST        /anyplace/floortiles/:buid/:floor_number                                controllers.MappingController.serveFloorPlanTilesZipLink( buid: String, floor_number: String )
-GET         /anyplace/floortiles/:buid/:floor_number/*file                          controllers.MappingController.serveFloorPlanTilesStatic( buid: String, floor_number: String, file: String )
+POST        /anyplace/floortiles/:buid/:floor_number                                controllers.MapFloorplanController.getZipLink( buid: String, floor_number: String )
+GET         /anyplace/floortiles/:buid/:floor_number/*file                          controllers.MapFloorplanController.getStaticTiles( buid: String, floor_number: String, file: String )
 # USED BY: Android Navigator:
-POST        /anyplace/floortiles/zip/:buid/:floor_number                            controllers.MappingController.serveFloorPlanTilesZip( buid: String, floor_number: String )
+POST        /anyplace/floortiles/zip/:buid/:floor_number                            controllers.MapFloorplanController.getTilesZip( buid: String, floor_number: String )
 # USED BY: Web apps
-POST        /anyplace/radiomaps_frozen/:building/:floor/:filename                   controllers.PositioningController.serveFrozenRadioMap(building: String, floor: String, filename: String)
-POST        /anyplace/radiomaps/:radio_folder/:filename                             controllers.PositioningController.serveRadioMap( radio_folder: String, filename: String )
+POST        /anyplace/radiomaps_frozen/:building/:floor/:filename                   controllers.RadiomapController.getFrozen(building: String, floor: String, filename: String)
+POST        /anyplace/radiomaps/:radio_folder/:filename                             controllers.RadiomapController.get( radio_folder: String, filename: String )
 
 ## OTHER:
 GET /favicon.ico                                                                    controllers.Assets.at(path="/public/images", file="favicon.png")
