+++ server/database/migration/test/README.md
@@ -1,0 +1,3 @@
+This folder contains some experiments and testing written in bash during `nneof` BSc thesis and internship.
+
+These are probably outdated and are here just for reference.
